+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "30.0.0"
 
     defaultConfig {
-        applicationId "com.example.pokemontypecalculator"
+        applicationId "com.perozzi_package.pokemontypecalculator"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
