--- app/src/main/java/com/example/pokemontypecalculator/TypeGridView.kt
@@ -1,17 +1,0 @@
-package com.example.pokemontypecalculator
-
-class TypeGridView {
-
-    var iconsInGridView: Int?
-    var textInGridView: Double?
-    var backgroundColorInGridView: Int?
-    var textColorInGridView: Int?
-
-    constructor(iconsInGridView: Int?, textInGridView: Double?, backgroundColorInGridView: Int?, textColorInGridView: Int?) {
-        this.iconsInGridView = iconsInGridView
-        this.textInGridView = textInGridView
-        this.backgroundColorInGridView = backgroundColorInGridView
-        this.textColorInGridView = textColorInGridView
-    }
-
-}
