+++ app/src/main/java/com/perozzi_package/pokemontypecalculator/TypeGridView.kt
@@ -1,0 +1,17 @@
+package com.perozzi_package.pokemontypecalculator
+
+class TypeGridView {
+
+    var iconsInGridView: Int?
+    var textInGridView: Double?
+    var backgroundColorInGridView: Int?
+    var textColorInGridView: Int?
+
+    constructor(iconsInGridView: Int?, textInGridView: Double?, backgroundColorInGridView: Int?, textColorInGridView: Int?) {
+        this.iconsInGridView = iconsInGridView
+        this.textInGridView = textInGridView
+        this.backgroundColorInGridView = backgroundColorInGridView
+        this.textColorInGridView = textColorInGridView
+    }
+
+}
